default config moved to the config controller
