+++ server/controllers/duels.js
@@ -2,13 +2,8 @@ import { queryPromise } from "../db/database.js";
 import HttpResponseWrapper from "../global/HttpResponseWrapper.js";
 import Logger, { addErrorTitle } from "../global/Logger.js";
 
+import { fetchConfigFromDB, DEFAULT_CONFIG } from "./config.js";
 import { createGeneratorOfType, NotEnoughDataError, getAllQuestionTypes } from "./question.js";
-
-export const DEFAULT_CONFIG = {
-  ROUNDS_PER_DUEL: 5,
-  QUESTIONS_PER_ROUNDS: 5,
-  QUESTION_TIMER_DURATION: 10,
-};
 
 /**
  *
@@ -52,7 +47,7 @@ function create(req, _res) {
       if (!yes) {
         return res.sendUsageError(404, "Opponent not found");
       }
-      getConfig()
+      fetchConfigFromDB()
         .then((config) =>
           createRounds(config)
             .then((rounds) =>
@@ -337,6 +332,7 @@ function createDuelInDatabase(player1, player2, rounds) {
 
 /**
  * Create all rounds of a duel
+ * @param {object} config The configuration object
  * This function can be mocked : @see _initMockedDuelRounds
  * @return {Promise<object[][]|NotEnoughDataError>}
  */
@@ -375,6 +371,7 @@ function createRounds(config) {
 /**
  * Create a round of a given question type
  * @param {number} type The question type
+ * @param {object} config The configuration object
  * @returns {Promise<object[]>} The list of questions
  */
 function createRound(type, config) {
@@ -627,28 +624,4 @@ function computeScores(duel) {
     },
     { user: 0, opponent: 0 }
   );
-}
-
-/**
- * Get the duel configuration in the database
- * @returns {Promise<object>} The config object
- */
-function getConfig() {
-  return new Promise((resolve, reject) => {
-    const sql = "SELECT * FROM server_informations WHERE server_informations.key LIKE 'config%';";
-
-    queryPromise(sql)
-      .then((res) => {
-        const { QUESTIONS_PER_ROUNDS, QUESTION_TIMER_DURATION, ROUNDS_PER_DUEL } = DEFAULT_CONFIG;
-        const getValue = (key) => Number(res.find((row) => row.key === key)?.value);
-
-        resolve({
-          questionsPerRounds: getValue("config_duel_questions_per_round") || QUESTIONS_PER_ROUNDS,
-          roundsPerDuel: getValue("config_duel_rounds_per_duel") || ROUNDS_PER_DUEL,
-          questionTimerDuration:
-            getValue("config_question_timer_duration") || QUESTION_TIMER_DURATION,
-        });
-      })
-      .catch(reject);
-  });
 }
