+++ server/test/index.test.js
@@ -97,11 +97,7 @@ function resetConfig() {
   return new Promise((resolve, reject) => {
     const sql = `DELETE FROM server_informations WHERE server_informations.key LIKE 'config%'; `;
     queryPromise(sql)
-      .then(() =>
-        insertData("config.sql")
-          .then(() => resolve())
-          .catch(reject)
-      )
+      .then(() => resolve())
       .catch(reject);
   });
 }
